Cek user terdaftar dan Redesign login dan daftar

+++ admin-kota_hapus.php
@@ -2,6 +2,13 @@
    
     require_once('konfigurasi.php');
 
+	$loggedin = isset($_SESSION['sedang_login']) && $_SESSION['sedang_login'] == 1;
+
+	if ( !$loggedin ) {
+	    header("Location: login.php");
+	    exit();
+	}
+
     $deletesql = "DELETE FROM kota    
     WHERE id = '{$_GET['kotaid']}'";
     $koneksidb->exec($deletesql);
